Updated patched poi jar to 3.8-beta4 and changed the patching to be lighter - just adding accessor methods, and then doing the rest within RightField itself. This should make future patching simpler. The patching cannot be avoided altogether (in the current poi version) as it is required to read and clear the DataValidityTable.
failed to completely remove the need for a patched poi.jar due to required access to the DataValidityTable. However, made patch lighter and just requiring an accessor to the DataValidityTable and DVRecord._formula1 and DVRecord._formula2, with the rest moved over to RightField. Still need to handle cleanDataValidations

added poi changes to enable DataValidationTable.clear()

added patch file for poi

isolated interaction with patched poi into a singleton class PatchedPoi